% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcequotes-package.R
\docType{package}
\name{forcequotes}
\alias{forcequotes}
\alias{forcequotes-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/forcequotes}
\item BugReports: \url{https://gitlab.com/hrbrmstr/forcequotes/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
