% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-quote.R
\name{random_quote}
\alias{random_quote}
\title{Return a Random Star Wars Quote}
\usage{
random_quote(faction = NULL)
}
\arguments{
\item{faction}{if not \code{NULL} than a value of \code{0:4} to specify a faction}
}
\value{
data frame of a single quote
}
\description{
Return a Random Star Wars Quote
}
\examples{
random_quote()
random_quote(faction = 3)
}
\references{
\url{http://swquotesapi.digitaljedi.dk/index.html};
\url{http://swquotes.digitaljedi.dk/home}
}
