% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote-console.R
\name{quote_console}
\alias{quote_console}
\title{Get a random Star Wars quote and display it stylishly to the console}
\usage{
quote_console(faction = NULL)
}
\arguments{
\item{faction}{faction (see \code{\link[=random_quote]{random_quote()}})}
}
\description{
Get a random Star Wars quote and display it stylishly to the console
}
